requirements.txt
matplotlib==3.10.0
numpy==2.2.1
Pillow==11.1.0
scikit_learn==1.6.1
scipy==1.15.0
torch==2.4.1+cu118
torchvision==0.19.1+cu118
tqdm==4.67.1
